+++ tests/portalbackend/PortalBackend.DBAccess.Tests/ServiceRepositoryTests.cs
@@ -39,17 +39,16 @@ namespace Org.CatenaX.Ng.Portal.Backend.PortalBackend.DBAccess.Tests;
 /// </summary>
 public class ServiceRepositoryTests : IAssemblyFixture<TestDbFixture>
 {
-    private readonly IFixture _fixture;
     private readonly TestDbFixture _dbTestDbFixture;
     private readonly Guid _offerId = new("99C5FD12-8085-4DE2-ABFD-215E1EE4BAA5");
 
     public ServiceRepositoryTests(TestDbFixture testDbFixture)
     {
-        _fixture = new Fixture().Customize(new AutoFakeItEasyCustomization { ConfigureMembers = true });
-        _fixture.Behaviors.OfType<ThrowingRecursionBehavior>().ToList()
-            .ForEach(b => _fixture.Behaviors.Remove(b));
+        var fixture = new Fixture().Customize(new AutoFakeItEasyCustomization { ConfigureMembers = true });
+        fixture.Behaviors.OfType<ThrowingRecursionBehavior>().ToList()
+            .ForEach(b => fixture.Behaviors.Remove(b));
 
-        _fixture.Behaviors.Add(new OmitOnRecursionBehavior());
+        fixture.Behaviors.Add(new OmitOnRecursionBehavior());
         _dbTestDbFixture = testDbFixture;
     }
 
@@ -164,8 +163,7 @@ public class ServiceRepositoryTests : IAssemblyFixture<TestDbFixture>
     private async Task<(OfferRepository, PortalDbContext)> CreateSut()
     {
         var context = await _dbTestDbFixture.GetPortalDbContext().ConfigureAwait(false);
-        _fixture.Inject(context);
-        var sut = _fixture.Create<OfferRepository>();
+        var sut = new OfferRepository(context);
         return (sut, context);
     }
 }