+++ tests/shared/Tests.Shared/TestSeeds/OfferData.cs
@@ -30,6 +30,7 @@ public static class OfferData
         new Offer(new Guid("99C5FD12-8085-4DE2-ABFD-215E1EE4BAA4"), "Catena X", DateTimeOffset.UtcNow, OfferTypeId.APP)
         {
             OfferStatusId = OfferStatusId.ACTIVE,
+            ProviderCompanyId = new Guid("2dc4249f-b5ca-4d42-bef1-7a7a950a4f87")
         },
         new Offer(new Guid("99C5FD12-8085-4DE2-ABFD-215E1EE4BAA5"), "Catena X", DateTimeOffset.UtcNow, OfferTypeId.SERVICE)
         {