+++ CHANGELOG.md
@@ -5,6 +5,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 ### In Progess
 
 * Feature - Service customer notification & email for service activation
+* Feature - Service provider notification & email for service subscriptions
 * Feature - User Management: Modification of Roles of one specific user under an app
 * Feature - Managed/Own connector registration endpoint refactoring
 * Feature - Enhance Logging for external Service calls